First pass at a main dispatcher which re-dispatches to a set of listeners.
Each listener watches it's own tube.

Right now there is just one listener: the debug listener. But it
can be used as the superclass for future listeners.

The main dispatcher is started via a twanager command, it then
forks off the additional lisetners as needed. When the main
dispatcher is terminated, the children are reaped.